Ron comment: rename operator.go and merge reconciler.go into it.
